Setup deploy keys to allow r/w on the repo
Added a deploy key to allow the travis CI
pipeline to push on the repository. Had to encode
the private key to base64 so that travis CI doesn't
bother the pipeline when escaping the environment variables.
Added the decoding in the stage `before_install`.
Removed the verbose option for the ssh-add as it was only
causing problems when running in the pipeline.
Changed https to ssh in the git url